Rename srtm to htool_srtm.
For consistency with other htool commands.
