+++ examples/meson.build
@@ -38,8 +38,8 @@ executable(
     'htool_spi_proxy.c',
     'htool_statistics.c',
     'htool_usb.c',
-    'srtm.h',
-    'srtm.c',
+    'htool_srtm.h',
+    'htool_srtm.c',
     git_version_h],
   dependencies: [libusb],
   link_with: link_with,
